Cloud_IoT_DIY_cloud/requirements.txt
attrs==22.2.0
aws-cdk-lib==2.63.2
aws-cdk.asset-awscli-v1==2.2.57
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.46
aws-cdk.aws-iot-actions-alpha==2.63.2a0
aws-cdk.aws-iot-alpha==2.63.2a0
aws-cdk.aws-iotevents-alpha==2.63.2a0
aws-cdk.aws-kinesisfirehose-alpha==2.63.2a0
boto3==1.26.74
botocore==1.29.74
cattrs==22.2.0
constructs==10.1.231
exceptiongroup==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
jsii==1.74.0
MarkupSafe==2.1.2
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
